requirements.txt
fastapi==0.104.1
mysql-connector-python==8.2.0
passlib==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
uvicorn==0.23.2